integrations/dspy/pyproject.toml
[project]
name = "databricks-dspy"
version = "0.0.1"
description = "Databricks integration for DSPy"
authors = [
    { name="Databricks", email="agent-feedback@databricks.com" },
]
readme = "README.md"
license = { text="Apache-2.0" }
requires-python = ">=3.10"
dependencies = [
    "dspy>=2.6.27",
    "databricks-sdk>=0.58.0",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
  "src/databricks_dspy/*"
]

[tool.hatch.build.targets.wheel]
packages = ["src/databricks_dspy"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
  # isort
  "I",
  # bugbear rules
  "B",
  # remove unused imports
  "F401",
  # bare except statements
  "E722",
  # print statements
  "T201",
  "T203",
  # misuse of typing.TYPE_CHECKING
  "TCH004",
  # import rules
  "TID251",
  # undefined-local-with-import-star
  "F403",
]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 88
